requirements.txt
asgiref
certifi
cffi
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==3.1.4
django-allauth
django-appconf
django-bootstrap-form
django-crispy-forms
django-etc==1.4.0
django-hitcount==1.3.5
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
pep8==1.7.1
pillow==10.2.0
pinax-messages
pinax-templates
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.0
whitenoise==6.6.0
